+++ karma.conf.js
@@ -67,6 +67,33 @@ module.exports = function(config) {
         base: 'SauceLabs',
         browserName: 'safari',
         platform: 'OS X 10.11'
+      },
+      slOpera12Win7: {
+        base: 'SauceLabs',
+        browserName: 'opera',
+        platform: 'Windows 7',
+        version: '12.12'
+      },
+      slIphone: {
+        base: 'SauceLabs',
+        browserName: 'iphone',
+        platform: 'OS X 10.10',
+        version: '8.2',
+        deviceName: 'iPhone Simulator'
+      },
+      slIpad: {
+        base: 'SauceLabs',
+        browserName: 'iphone',
+        platform: 'OS X 10.10',
+        version: '8.2',
+        deviceName: 'iPad Simulator'
+      },
+      slAndroid: {
+        base: 'SauceLabs',
+        browserName: 'android',
+        platform: 'Linux',
+        version: '5.1',
+        deviceName: 'Android Emulator'
       }
     };
     browsers = Object.keys(customLaunchers);
